requirements.txt
feedparser
PyYAML
requests
Markdown
Jinja2
